+++ src/melee/ft/ftdrawcommon.h
@@ -6,7 +6,7 @@
 #include <dolphin/mtx/types.h>
 #include <baselib/gobj.h>
 
-void ftDrawCommmon_80080E18(HSD_GObj*, long);
+void ftDrawCommmon_80080E18(HSD_GObj*, int);
 MtxPtr ftDrawCommmon_8008051C(HSD_GObj*, MtxPtr);
 
 void ftDrawCommmon_800805C8(HSD_GObj* gobj, s32 arg1, s32 arg2);
